removed NativeSecp256k1 unnecesary class
